# target/classes/api.yaml
openapi: 3.0.2
info:
  title: Swagger Saga Pattern Order Service - OpenAPI 3.0
  description: |-
    OpenAPI 3.0 specification.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: viniciusnrgarcia@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 0.0.1
servers:
  - url: /api/v1

tags:
  - name: order
    description: Order operations
    externalDocs:
      description: More information
      url: 'http://swagger.io'

paths:
  /order:
    post:
      tags:
        - order
      summary: Add a new order
      description: Add a new order
      operationId: addOrder
      requestBody:
        description: Create a new Order
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '405':
          description: Invalid input
        '500':
          description: Internal server error
    put:
      tags:
        - order
      summary: Update an existent Order
      description: Update an existent Order by Id
      operationId: updateOrder
      requestBody:
        description: Update an existent Order
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Order not found
        '405':
          description: Validation exception
        '500':
          description: Internal server error

  '/order/{orderId}':
    delete:
      tags:
        - order
      summary: Deletes a Order
      description: Delete an existent Order by Id
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          description: Order id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '400':
          description: Invalid Order Id
        '500':
          description: Internal server error


components:
  schemas:
    Order:
      type: object
      properties:
        orderType:
          type: integer
          minimum: 1
        orderState:
          type: integer
          minimum: 1
        orderDescription:
          type: string
        orderValue:
          type: number
        productId:
          type: integer
          format: int64
        productItems:
          type: integer
          minimum: 1
        productDescription:
          type: string
        customerId:
          type: string
          format: uuid

    ModelApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string